Moved createIndexes method to DAO
