+++ rest-server/core/src/main/scala/com/actionml/core/store/backends/MongoDao.scala
@@ -25,6 +25,25 @@ import org.mongodb.scala.bson.ObjectId
 import org.mongodb.scala.bson.collection.immutable.Document
 import org.mongodb.scala.bson.conversions.Bson
 import org.mongodb.scala.model.{Filters, Sorts}
+import java.time.{Instant, OffsetDateTime, ZoneOffset}
+import java.util.concurrent.TimeUnit
+
+import com.actionml.core.store.indexes.annotations.Indexed
+import com.actionml.core.store.{DAO, Ordering, Store}
+import com.mongodb.client.model.IndexOptions
+import com.typesafe.scalalogging.LazyLogging
+import org.bson.codecs.configuration.{CodecProvider, CodecRegistries}
+import org.bson.codecs.{Codec, DecoderContext, EncoderContext}
+import org.bson.{BsonReader, BsonWriter}
+import org.mongodb.scala.bson.collection.immutable.Document
+import org.mongodb.scala.model.IndexModel
+import org.mongodb.scala.{MongoClient, MongoCollection, MongoDatabase}
+
+import scala.concurrent.duration._
+import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.reflect.ClassTag
+import scala.reflect.runtime.universe._
+
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.reflect.ClassTag
@@ -32,6 +51,7 @@ import scala.reflect.ClassTag
 
 class MongoDao[T](val collection: MongoCollection[T])(implicit ct: ClassTag[T]) extends DAO[T] with LazyLogging {
   import DaoQuery.syntax._
+  import scala.concurrent.ExecutionContext.Implicits.global
 
   override def name = collection.namespace.getFullName
   override def dbName = collection.namespace.getDatabaseName
@@ -135,6 +155,34 @@ class MongoDao[T](val collection: MongoCollection[T])(implicit ct: ClassTag[T]) 
         logger.error(s"Can't removeMany from collection ${collection.namespace} with filter $filter")
         Future.failed(new RuntimeException(s"Can't removeMany from collection ${collection.namespace} with filter $filter"))
     }
+  }
+
+  // compare actual and required indexes and create missing ones
+  override def createIndexes(indexesDescription: Seq[(String, Indexed)]): Future[Unit] = {
+    // get information about indexes from mongo db
+    def getActualIndexesInfo(col: MongoCollection[T]): Future[Seq[(String, Indexed)]] = {
+      col.listIndexes().map { i =>
+        val keyInfo = i.get("key").get.asDocument()
+        val iName = keyInfo.getFirstKey
+        (iName, Indexed())
+      }.toFuture
+    }
+    getActualIndexesInfo(collection)
+      .map(actualIndexesInfo => indexesDescription diff actualIndexesInfo.toList)
+      .flatMap { absentIndexes =>
+        val newIndexes = absentIndexes.map { case (iName, Indexed(iOrder, ttl)) =>
+          val options = new IndexOptions()
+          if (ttl.isFinite()) options.expireAfter(ttl.toMillis, TimeUnit.MILLISECONDS)
+          IndexModel(Document(iName -> 1), options)
+        }
+        if (newIndexes.nonEmpty) {
+          logger.info(s"Creating indexes ${newIndexes.map(i => s"${i.getKeys} - ${i.getOptions}")} for collection ${collection.namespace.getFullName}")
+          collection.createIndexes(newIndexes).toFuture
+        } else Future.successful(())
+      }.recover {
+      case e: Exception =>
+        logger.error(s"Can't create indexes for ${collection.namespace.getFullName}", e)
+    }.map(_ => ())
   }
